refactor: align kda Triton module layout.
